+++ packages/cli/src/extract-experimental/constants.ts
@@ -1,2 +1,24 @@
 export const ENTRY_NAME_PH = "{entryName}"
 export const DEFAULT_TEMPLATE_NAME = "messages"
+
+export const DEFAULT_EXCLUDE_EXTENSIONS = [
+  "ico",
+  "pot",
+  "po",
+  "webp",
+  "xliff",
+  "woff2",
+  "woff",
+  "eot",
+  "gif",
+  "otf",
+  "ttf",
+  "mp4",
+  "svg",
+  "png",
+  "css",
+  "sass",
+  "scss",
+  "less",
+  "jpg",
+]