+++ packages/cli/src/extract-experimental/workers/extractWorker.ts
@@ -1,45 +1,26 @@
-import { getConfig, LinguiConfigNormalized } from "@lingui/conf"
-import { extractFromBundleAndWrite } from "../extractFromBundleAndWrite.js"
-import { FormatterWrapper, getFormat } from "../../api/formats/index.js"
+import {
+  ExtractedMessage,
+  getConfig,
+  LinguiConfigNormalized,
+} from "@lingui/conf"
+import { extractFromChunk } from "../extractFromChunk.js"
 
 export type ExtractWorkerFunction = typeof extractWorker
 
 let linguiConfig: LinguiConfigNormalized | undefined
-let format: FormatterWrapper | undefined
 
 const extractWorker = async (
   linguiConfigPath: string,
-  entryPoint: string,
   bundleFile: string,
-  outputPattern: string,
-  template: boolean,
-  locales: string[],
-  clean: boolean,
-  overwrite: boolean,
-) => {
+): Promise<{ success: boolean; messages: ExtractedMessage[] }> => {
   if (!linguiConfig) {
-    // initialize config once per worker, speed up workers follow execution
     linguiConfig = getConfig({
       configPath: linguiConfigPath,
       skipValidation: true,
     })
   }
 
-  if (!format) {
-    format = await getFormat(linguiConfig.format, linguiConfig.sourceLocale)
-  }
-
-  return await extractFromBundleAndWrite({
-    entryPoint,
-    bundleFile,
-    outputPattern,
-    format,
-    linguiConfig,
-    locales,
-    overwrite,
-    clean,
-    template,
-  })
+  return await extractFromChunk(bundleFile, linguiConfig)
 }
 
 export { extractWorker }