+++ packages/cli/test/extractor-experimental/expected-rolldown/about.page.pl.js
@@ -1,0 +1,1 @@
+/*eslint-disable*/module.exports={messages:JSON.parse("{\"LGGfGX\":[\"translation: header message\"],\"RNRWaX\":[\"shared message\"],\"VmkjGB\":[\"Green\"],\"nPi9F1\":[\"this should be marked as obsolete\"],\"u5PTM8\":[\"about page message\"]}")};